school_registry/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.0.1
python-dotenv==1.0.0
Flask-Migrate==4.0.5
SQLAlchemy==1.4.49
psycopg2-binary
Werkzeug==2.3.7
email-validator==2.1.0
pandas
openpyxl==3.0.10
faker